Only run cronjobs on job servers

+++ lib/tasks/exports.rake
@@ -10,6 +10,11 @@ namespace :exports do
 
   desc "Remove expired exports."
   task expire: :environment do
+    unless ENV["JOB_SERVER"] == "true"
+      puts "SKIP: Not running on job server."
+      next
+    end
+
     Export.expired.destroy_all
   end
 end
